Trim out the old code used for running progress checks on a cron job on my 10-year-old raspberry pi
