+++ internal/config/config.go
@@ -1,0 +1,6 @@
+package config
+
+const (
+	AWSRegion              = "us-west-2"
+	HistoryDynamoTableName = "storm-charts"
+)
