+++ internal/progress/errorPage.go
@@ -1,0 +1,8 @@
+package progress
+
+import (
+	_ "embed"
+)
+
+//go:embed error-page.html
+var ErrorPageContent []byte
